+++ README.md
@@ -223,7 +223,7 @@ Alternatively, you can run H2O LLM Studio GUI by using our self-hosted Docker im
 You can also use H2O LLM Studio with the command line interface (CLI) and specify the configuration .yaml file that contains all the experiment parameters. To finetune using H2O LLM Studio with CLI, activate the pipenv environment by running `make shell`, and then use the following command:
 
 ```bash
-python train.py -Y {path_to_config_yaml_file}
+python llm_studio.train.py -Y {path_to_config_yaml_file}
 ```
 
 To run on multiple GPUs in DDP mode, run the following command:
@@ -237,7 +237,7 @@ By default, the framework will run on the first `k` GPUs. If you want to specify
 To start an interactive chat with your trained model, use the following command:
 
 ```bash
-python prompt.py -e {experiment_name}
+python llm_studio.prompt.py -e {experiment_name}
 ```
 
 where `experiment_name` is the output folder of the experiment you want to chat with (see configuration).
@@ -248,7 +248,7 @@ To publish the model to Hugging Face, use the following command:
 ```bash
 make shell 
 
-python publish_to_hugging_face.py -p {path_to_experiment} -d {device} -a {api_key} -u {user_id} -m {model_name} -s {safe_serialization}
+python llm_studio.publish_to_hugging_face.py -p {path_to_experiment} -d {device} -a {api_key} -u {user_id} -m {model_name} -s {safe_serialization}
 ```
 
 `path_to_experiment` is the output folder of the experiment.
