+++ entrypoint.sh
@@ -5,4 +5,4 @@ nvidia-smi
 
 echo "Starting H2O LLM Studio..."
 
-wave run --no-reload app
+wave run --no-reload llm_studio.app
